First implementation of the script
The script now works, recursing into every directory from the given path and setting the cover art on the parent folder of the cover image as well as all audio files in the folder of the cover art. Currently the files supported are hardcoded in like 39. Supported files are: MP3, FLAC, M4A, OGG Vorbis, WAV, AIFF, WMA, and ALAC